Exclude Error Prone as a default
While well-recommended, Error Prone may be enough outside most experience
that including it in the build (even disabled) would be distracting. It
is kept, however, as advice in the README.

Note also: There are two common Gradle plugins for Error Prone. This
project had been using one of them; the author of that one is a
contributor to the other one; the installation page for Error Prone
mentions the other one.